move FitData to util
