+++ src/ImageFitting/ImageFitter.h
@@ -20,32 +20,11 @@
 #include <imageanalysis/ImageTypedefs.h>
 
 #include "ImageFitting/InitialValueCalculator.h"
+#include "ImageFitting/Util.h"
 #include "ImageGenerators/ImageGenerator.h"
 #include "Logger/Logger.h"
 
 namespace carta {
-
-/** @brief Data structure for storing fitting-related data. */
-struct FitData {
-    /** @brief Pointer to the image data. */
-    float* data;
-    /** @brief The width of the image. */
-    size_t width;
-    /** @brief Number of pixels. */
-    size_t n;
-    /** @brief Number of pixels excluding nan pixels. */
-    size_t n_notnan;
-    /** @brief X-axis offset from the fitting region to the entire image. */
-    size_t offset_x;
-    /** @brief Y-axis offset from the fitting region to the entire image. */
-    size_t offset_y;
-    /** @brief Indexes of the Gaussian parameters in the fittig parameters. */
-    std::vector<int> fit_values_indexes;
-    /** @brief Initial fitting parameters. */
-    std::vector<double> initial_values;
-    /** @brief Whether to stop the fitting process. */
-    bool stop_fitting;
-};
 
 /** @brief Data structure for storing status of the fitting result. */
 struct FitStatus {
